+++ cmd/qa/main.go
@@ -1,7 +1,0 @@
-//go:build hydreq_removed
-// +build hydreq_removed
-
-// Deprecated entrypoint: use `hydreq` instead.
-package main
-
-func main() {}